services/auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Authentication service for Agora Electoral"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlmodel = "^0.0.20"
psycopg2-binary = "^2.9.9" # O asyncpg para uso asíncrono puro

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
pythonpath = ["app"]
